We have made some changes in the code as well as we have added graph plotter in every program so that we can also visualise the result

+++ Text_cleaner.py
@@ -1,10 +1,11 @@
+import os
 # This code removes the problems which were occuring before and now we have better clean and parsed data
 replace = ", ‚ ' ; : / ™ [ ] { } ( ) * - + & ! @ # $ % ^ _ = ` ‘ “ ~"
 
 if __name__ == "__main__":
 	i = 0
 	# change the value of this variable with the number of files that have been parsed
-	max1 = 11
+	max1 = 43
 	with open("C:/Users/LENOVO/Desktop/Internship/To_download_links.txt",'r') as t:
 		x = t.read()
 		for link in x.split():
@@ -14,17 +15,23 @@ if __name__ == "__main__":
 				print(link)
 				try:
 					s = 0
-					with open('C:/Users/LENOVO/Desktop/Internship/Newpaper_Cleaned/'+link+'.txt','rb') as f:
-						x = f.read()
-						x = x.decode("utf-8")
-						t = str(x)
-						for r in replace.split():
-							t = t.replace(r,'')
-						t = t.replace("\n"," ")
-						t = t.replace("-","")
-						s = t
-					with open('C:/Users/LENOVO/Desktop/Internship/Better_cleaned/'+link+'.txt','wb') as f:
-						f.write(s.encode("utf-8"))
-					f.close()
+					# print(link)
+					if os.path.isfile("C:/Users/LENOVO/Desktop/Internship/Better_cleaned/"+link+'.txt'):
+						print("The file has already been parsed")
+						print()
+
+					else:
+						with open('C:/Users/LENOVO/Desktop/Internship/Newpaper_Cleaned/'+link+'.txt','rb') as f:
+							x = f.read()
+							x = x.decode("utf-8")
+							t = str(x)
+							for r in replace.split():
+								t = t.replace(r,'')
+							t = t.replace("\n"," ")
+							t = t.replace("-","")
+							s = t
+						with open('C:/Users/LENOVO/Desktop/Internship/Better_cleaned/'+link+'.txt','wb') as f:
+							f.write(s.encode("utf-8"))
+						f.close()
 				except:
 					pass
